+++ bee-ledger/src/workers/pruning/batch.rs
@@ -37,7 +37,7 @@ pub struct Edge {
     pub to_child: MessageId,
 }
 
-pub async fn prune_confirmed_data<S: StorageBackend>(
+pub fn prune_confirmed_data<S: StorageBackend>(
     tangle: &Tangle<S>,
     storage: &S,
     batch: &mut S::Batch,
@@ -207,7 +207,7 @@ pub async fn prune_confirmed_data<S: StorageBackend>(
     Ok((new_seps, metrics))
 }
 
-pub async fn prune_unconfirmed_data<S: StorageBackend>(
+pub fn prune_unconfirmed_data<S: StorageBackend>(
     storage: &S,
     batch: &mut S::Batch,
     prune_index: MilestoneIndex,
@@ -318,7 +318,7 @@ pub async fn prune_unconfirmed_data<S: StorageBackend>(
     Ok(metrics)
 }
 
-pub async fn prune_milestone_data<S: StorageBackend>(
+pub fn prune_milestone_data<S: StorageBackend>(
     storage: &S,
     batch: &mut S::Batch,
     prune_index: MilestoneIndex,
@@ -326,12 +326,12 @@ pub async fn prune_milestone_data<S: StorageBackend>(
 ) -> Result<MilestoneDataPruningMetrics, Error> {
     let mut metrics = MilestoneDataPruningMetrics::default();
 
-    prune_milestone(storage, batch, prune_index).await?;
+    prune_milestone(storage, batch, prune_index)?;
 
-    prune_output_diff(storage, batch, prune_index).await?;
+    prune_output_diff(storage, batch, prune_index)?;
 
     if should_prune_receipts {
-        metrics.receipts = prune_receipts(storage, batch, prune_index).await?;
+        metrics.receipts = prune_receipts(storage, batch, prune_index)?;
     }
 
     Ok(metrics)
@@ -359,22 +359,14 @@ fn prune_edge<S: StorageBackend>(
     Ok(())
 }
 
-async fn prune_milestone<S: StorageBackend>(
-    storage: &S,
-    batch: &mut S::Batch,
-    index: MilestoneIndex,
-) -> Result<(), Error> {
+fn prune_milestone<S: StorageBackend>(storage: &S, batch: &mut S::Batch, index: MilestoneIndex) -> Result<(), Error> {
     Batch::<MilestoneIndex, Milestone>::batch_delete(storage, batch, &index)
         .map_err(|e| Error::Storage(Box::new(e)))?;
 
     Ok(())
 }
 
-async fn prune_output_diff<S: StorageBackend>(
-    storage: &S,
-    batch: &mut S::Batch,
-    index: MilestoneIndex,
-) -> Result<(), Error> {
+fn prune_output_diff<S: StorageBackend>(storage: &S, batch: &mut S::Batch, index: MilestoneIndex) -> Result<(), Error> {
     if let Some(output_diff) =
         Fetch::<MilestoneIndex, OutputDiff>::fetch(storage, &index).map_err(|e| Error::Storage(Box::new(e)))?
     {
@@ -396,11 +388,7 @@ async fn prune_output_diff<S: StorageBackend>(
     Ok(())
 }
 
-async fn prune_receipts<S: StorageBackend>(
-    storage: &S,
-    batch: &mut S::Batch,
-    index: MilestoneIndex,
-) -> Result<usize, Error> {
+fn prune_receipts<S: StorageBackend>(storage: &S, batch: &mut S::Batch, index: MilestoneIndex) -> Result<usize, Error> {
     let receipts = Fetch::<MilestoneIndex, Vec<Receipt>>::fetch(storage, &index)
         .map_err(|e| Error::Storage(Box::new(e)))?
         // Fine since Fetch of a Vec<_> always returns Some(Vec<_>).
@@ -419,11 +407,7 @@ async fn prune_receipts<S: StorageBackend>(
 
 // TODO: consider using this instead of 'truncate'
 #[allow(dead_code)]
-async fn prune_seps<S: StorageBackend>(
-    storage: &S,
-    batch: &mut S::Batch,
-    seps: &[SolidEntryPoint],
-) -> Result<usize, Error> {
+fn prune_seps<S: StorageBackend>(storage: &S, batch: &mut S::Batch, seps: &[SolidEntryPoint]) -> Result<usize, Error> {
     let mut num = 0;
     for sep in seps {
         Batch::<SolidEntryPoint, MilestoneIndex>::batch_delete(storage, batch, sep)