+++ bee-storage/bee-storage-rocksdb/src/access/batch.rs
@@ -25,6 +25,7 @@ use crate::{
 
 #[derive(Default)]
 pub struct StorageBatch {
+    should_lock: bool,
     inner: WriteBatch,
     key_buf: Vec<u8>,
     value_buf: Vec<u8>,
@@ -37,7 +38,12 @@ impl BatchBuilder for Storage {
         let mut write_options = WriteOptions::default();
         write_options.set_sync(false);
         write_options.disable_wal(!durability);
+
+        let guard = batch.should_lock.then(|| self.locks.message_id_to_metadata.read());
+
         self.inner.write_opt(batch.inner, &write_options)?;
+
+        drop(guard);
 
         Ok(())
     }
@@ -81,6 +87,8 @@ impl Batch<MessageId, MessageMetadata> for Storage {
         message_id: &MessageId,
         metadata: &MessageMetadata,
     ) -> Result<(), <Self as StorageBackend>::Error> {
+        batch.should_lock = true;
+
         batch.value_buf.clear();
         // Packing to bytes can't fail.
         metadata.pack(&mut batch.value_buf).unwrap();
@@ -97,6 +105,8 @@ impl Batch<MessageId, MessageMetadata> for Storage {
         batch: &mut Self::Batch,
         message_id: &MessageId,
     ) -> Result<(), <Self as StorageBackend>::Error> {
+        batch.should_lock = true;
+
         batch
             .inner
             .delete_cf(self.cf_handle(CF_MESSAGE_ID_TO_METADATA)?, message_id);